+++ clients/web/src/components/guests/GuestVisitActivityTab.tsx
@@ -1,201 +1,16 @@
 import { useState } from "react";
-import {
-  CalendarDays,
-  ChevronRight,
-  Clock4,
-  FlagIcon,
-  UsersRound,
-} from "lucide-react";
+import { ChevronRight } from "lucide-react";
+import { ActiveBookingCard } from "./ActiveBookingCard";
 import { GuestBookingHistoryView } from "./GuestBookingHistoryView";
+import { PastBookingCard } from "./PastBookingCard";
+import { RequestCard } from "./RequestCard";
 import type { GuestRequest, Stay } from "@shared";
-import { cn } from "@/lib/utils";
-import { formatDate } from "@/utils/dates";
 
 type GuestVisitActivityTabProps = {
   currentStays: Array<Stay>;
   pastStays: Array<Stay>;
   requests: Array<GuestRequest>;
 };
-
-type Priority = "high" | "medium" | "low";
-
-const priorityConfig: Record<
-  Exclude<Priority, "low">,
-  { label: string; containerClass: string; contentClass: string }
-> = {
-  high: {
-    label: "High Priority",
-    containerClass: "bg-[#ffeded]",
-    contentClass: "text-[#a21313]",
-  },
-  medium: {
-    label: "Medium Priority",
-    containerClass: "bg-bg-orange",
-    contentClass: "text-text-orange",
-  },
-};
-
-function ActiveBookingCard({
-  stay,
-  compact,
-}: {
-  stay: Stay;
-  compact?: boolean;
-}) {
-  return (
-    <div
-      className={cn(
-        "flex flex-col gap-2 rounded-lg border border-primary bg-bg-selected p-4",
-        compact && "w-[231px] shrink-0",
-      )}
-    >
-      <div className="flex items-start justify-between">
-        <span className="text-xl font-bold text-primary">
-          Suite {stay.room_number}
-        </span>
-        {stay.group_size != null && (
-          <div className="flex items-center gap-1 text-primary">
-            <UsersRound className="size-[19px]" strokeWidth={1.5} />
-            <span className="text-xl font-bold">{stay.group_size}</span>
-          </div>
-        )}
-      </div>
-      {compact ? (
-        <div className="flex flex-col gap-1">
-          <div className="flex flex-col gap-1">
-            <span className="text-sm font-medium text-primary">Arrival:</span>
-            <div className="flex items-center gap-2 text-sm text-primary">
-              <CalendarDays className="size-3 shrink-0" strokeWidth={1.5} />
-              <span>{formatDate(stay.arrival_date)}</span>
-            </div>
-          </div>
-          <div className="flex flex-col gap-1">
-            <span className="text-sm font-medium text-primary">Departure:</span>
-            <div className="flex items-center gap-2 text-sm text-primary">
-              <CalendarDays className="size-3 shrink-0" strokeWidth={1.5} />
-              <span>{formatDate(stay.departure_date)}</span>
-            </div>
-          </div>
-        </div>
-      ) : (
-        <div className="flex gap-[72px]">
-          <div className="flex flex-1 flex-col gap-1">
-            <span className="text-sm font-medium text-primary">Arrival:</span>
-            <div className="flex items-center gap-2 text-sm text-primary">
-              <CalendarDays className="size-3 shrink-0" strokeWidth={1.5} />
-              <span>{formatDate(stay.arrival_date)}</span>
-            </div>
-          </div>
-          <div className="flex flex-1 flex-col gap-1">
-            <span className="text-sm font-medium text-primary">Departure:</span>
-            <div className="flex items-center gap-2 text-sm text-primary">
-              <CalendarDays className="size-3 shrink-0" strokeWidth={1.5} />
-              <span>{formatDate(stay.departure_date)}</span>
-            </div>
-          </div>
-        </div>
-      )}
-    </div>
-  );
-}
-
-function PastBookingCard({ stay }: { stay: Stay }) {
-  return (
-    <div className="flex items-center rounded-lg border border-text-subtle bg-bg-container p-4">
-      <div className="flex flex-1 flex-col gap-1">
-        <div className="flex items-center gap-2">
-          <span className="text-sm font-medium text-text-subtle">
-            Suite {stay.room_number}
-          </span>
-          {stay.group_size != null && (
-            <div className="flex items-center gap-0.5 text-text-subtle">
-              <UsersRound className="size-3" strokeWidth={1.5} />
-              <span className="text-sm font-medium">{stay.group_size}</span>
-            </div>
-          )}
-        </div>
-        <span className="text-sm text-text-subtle">
-          {formatDate(stay.arrival_date)} - {formatDate(stay.departure_date)}
-        </span>
-      </div>
-    </div>
-  );
-}
-
-function RequestCard({ req }: { req: GuestRequest }) {
-  const p = req.priority as Priority | undefined;
-  const tags = [req.request_category, req.request_type].filter(Boolean);
-
-  return (
-    <div className="flex flex-col gap-2 rounded-lg border border-text-subtle bg-white px-4 py-6">
-      {/* Priority badge + name */}
-      <div className="flex items-center gap-2">
-        {(p === "high" || p === "medium") && (
-          <div
-            className={cn(
-              "inline-flex items-center gap-1 rounded px-2 py-1",
-              priorityConfig[p].containerClass,
-            )}
-          >
-            <FlagIcon
-              className={cn("size-4", priorityConfig[p].contentClass)}
-              strokeWidth={2}
-            />
-            <span className={cn("text-xs", priorityConfig[p].contentClass)}>
-              {priorityConfig[p].label}
-            </span>
-          </div>
-        )}
-        <span className="text-sm font-semibold text-text-default">
-          {req.name}
-        </span>
-      </div>
-
-      {/* Room + timestamp */}
-      <div className="flex items-center gap-2 text-xs text-text-default">
-        {req.room_number != null && <span>Suite {req.room_number}</span>}
-        {req.created_at && (
-          <div className="flex items-center gap-0.5">
-            <Clock4 className="size-2.5" strokeWidth={1.5} />
-            <span>
-              {new Date(req.created_at).toLocaleTimeString([], {
-                hour: "numeric",
-                minute: "2-digit",
-              })}
-            </span>
-          </div>
-        )}
-      </div>
-
-      {/* Category/type tags */}
-      {tags.length > 0 && (
-        <div className="flex flex-wrap gap-2">
-          {tags.map((tag) => (
-            <span
-              key={tag}
-              className="inline-flex items-center rounded border border-primary bg-bg-selected px-2 py-1 text-xs text-primary"
-            >
-              {tag}
-            </span>
-          ))}
-        </div>
-      )}
-
-      {/* Description */}
-      {req.description && (
-        <p className="text-sm text-text-subtle">{req.description}</p>
-      )}
-
-      {/* Mark in Progress button */}
-      <button
-        type="button"
-        className="mt-1 w-full rounded bg-primary px-6 py-2 text-sm text-white hover:bg-primary-hover"
-      >
-        Mark in Progress
-      </button>
-    </div>
-  );
-}
 
 export function GuestVisitActivityTab({
   currentStays,
